Refactor image resizing into PeekabooImageResizer with async processing in Android
