+++ peekaboo-image-picker/src/androidMain/kotlin/com/preat/peekaboo/image/picker/ImagePickerLauncher.android.kt
@@ -15,26 +15,14 @@
  */
 package com.preat.peekaboo.image.picker
 
-import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.graphics.Canvas
-import android.graphics.ColorMatrix
-import android.graphics.ColorMatrixColorFilter
-import android.graphics.Matrix
-import android.graphics.Paint
-import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
-import androidx.exifinterface.media.ExifInterface
-import com.preat.peekaboo.image.picker.PeekabooBitmapCache.bitmapToByteArray
 import com.preat.peekaboo.image.picker.SelectionMode.Companion.INFINITY
 import kotlinx.coroutines.CoroutineScope
-import java.io.ByteArrayOutputStream
 
 @Composable
 actual fun rememberImagePickerLauncher(
@@ -47,6 +35,7 @@ actual fun rememberImagePickerLauncher(
     return when (selectionMode) {
         SelectionMode.Single ->
             pickSingleImage(
+                coroutineScope = scope,
                 selectionMode = selectionMode,
                 resizeOptions = resizeOptions,
                 filterOptions = filterOptions,
@@ -55,6 +44,7 @@ actual fun rememberImagePickerLauncher(
 
         is SelectionMode.Multiple ->
             pickMultipleImages(
+                coroutineScope = scope,
                 selectionMode = selectionMode,
                 resizeOptions = resizeOptions,
                 filterOptions = filterOptions,
@@ -65,6 +55,7 @@ actual fun rememberImagePickerLauncher(
 
 @Composable
 private fun pickSingleImage(
+    coroutineScope: CoroutineScope,
     selectionMode: SelectionMode,
     resizeOptions: ResizeOptions,
     filterOptions: FilterOptions,
@@ -77,16 +68,17 @@ private fun pickSingleImage(
             contract = ActivityResultContracts.PickVisualMedia(),
             onResult = { uri ->
                 uri?.let {
-                    val resizedImage =
-                        resizeImage(
-                            context = context,
-                            uri = uri,
-                            width = resizeOptions.width,
-                            height = resizeOptions.height,
-                            filterOptions = filterOptions,
-                        )
-                    if (resizedImage != null) {
-                        onResult(listOf(resizedImage))
+                    PeekabooImageResizer.resizeImageAsync(
+                        context = context,
+                        coroutineScope = coroutineScope,
+                        uri = uri,
+                        width = resizeOptions.width,
+                        height = resizeOptions.height,
+                        filterOptions = filterOptions,
+                    ) { resizedImage ->
+                        if (resizedImage != null) {
+                            onResult(listOf(resizedImage))
+                        }
                     }
                 }
                 imagePickerLauncher?.markPhotoPickerInactive()
@@ -108,6 +100,7 @@ private fun pickSingleImage(
 
 @Composable
 private fun pickMultipleImages(
+    coroutineScope: CoroutineScope,
     selectionMode: SelectionMode.Multiple,
     resizeOptions: ResizeOptions,
     filterOptions: FilterOptions,
@@ -126,18 +119,23 @@ private fun pickMultipleImages(
         rememberLauncherForActivityResult(
             contract = ActivityResultContracts.PickMultipleVisualMedia(maxSelection),
             onResult = { uriList ->
-                val imageBytesList =
-                    uriList.mapNotNull { uri ->
-                        resizeImage(
-                            context = context,
-                            uri = uri,
-                            width = resizeOptions.width,
-                            height = resizeOptions.height,
-                            filterOptions = filterOptions,
-                        )
+                val resizedImages = mutableListOf<ByteArray>()
+                uriList.forEach { uri ->
+                    PeekabooImageResizer.resizeImageAsync(
+                        context = context,
+                        coroutineScope = coroutineScope,
+                        uri = uri,
+                        width = resizeOptions.width,
+                        height = resizeOptions.height,
+                        filterOptions = filterOptions,
+                    ) { resizedImage ->
+                        resizedImage?.let {
+                            resizedImages.add(it)
+                            if (resizedImages.size == uriList.size) {
+                                onResult(resizedImages)
+                            }
+                        }
                     }
-                if (imageBytesList.isNotEmpty()) {
-                    onResult(imageBytesList)
                 }
                 imagePickerLauncher?.markPhotoPickerInactive()
             },
@@ -172,113 +170,4 @@ actual class ImagePickerLauncher actual constructor(
         isPhotoPickerActive = true
         onLaunch()
     }
-}
-
-private fun resizeImage(
-    context: Context,
-    uri: Uri,
-    width: Int,
-    height: Int,
-    filterOptions: FilterOptions,
-): ByteArray? {
-    val resizeCacheKey = "${uri}_w${width}_h$height"
-    val filterCacheKey = "${resizeCacheKey}_$filterOptions"
-
-    PeekabooBitmapCache.instance.get(filterCacheKey)?.let { cachedBitmap ->
-        return bitmapToByteArray(cachedBitmap)
-    }
-
-    val resizedBitmap =
-        PeekabooBitmapCache.instance.get(resizeCacheKey) ?: run {
-            context.contentResolver.openInputStream(uri)?.use { inputStream ->
-                val options = BitmapFactory.Options().apply { inJustDecodeBounds = true }
-                BitmapFactory.decodeStream(inputStream, null, options)
-
-                var inSampleSize = 1
-                while (options.outWidth / inSampleSize > width || options.outHeight / inSampleSize > height) {
-                    inSampleSize *= 2
-                }
-
-                options.inJustDecodeBounds = false
-                options.inSampleSize = inSampleSize
-
-                context.contentResolver.openInputStream(uri)?.use { scaledInputStream ->
-                    BitmapFactory.decodeStream(scaledInputStream, null, options)?.also { bitmap ->
-                        PeekabooBitmapCache.instance.put(resizeCacheKey, bitmap)
-                    }
-                }
-            }
-        }
-
-    resizedBitmap?.let {
-        val rotatedBitmap = rotateImageIfRequired(context, it, uri)
-        val filteredBitmap = applyFilter(rotatedBitmap, filterOptions)
-
-        ByteArrayOutputStream().use { byteArrayOutputStream ->
-            filteredBitmap.compress(Bitmap.CompressFormat.JPEG, 100, byteArrayOutputStream)
-            val byteArray = byteArrayOutputStream.toByteArray()
-            PeekabooBitmapCache.instance.put(filterCacheKey, filteredBitmap)
-            return byteArray
-        }
-    }
-
-    return null
-}
-
-private fun applyFilter(
-    originalBitmap: Bitmap,
-    filterOptions: FilterOptions,
-): Bitmap {
-    val colorMatrix = ColorMatrix()
-    when (filterOptions) {
-        FilterOptions.Default -> return originalBitmap
-        FilterOptions.GrayScale -> colorMatrix.setSaturation(0f)
-        FilterOptions.Sepia -> {
-            colorMatrix.setSaturation(0f)
-            val sepiaMatrix =
-                ColorMatrix().apply {
-                    setScale(1f, 0.95f, 0.82f, 1f)
-                }
-            colorMatrix.postConcat(sepiaMatrix)
-        }
-        FilterOptions.Invert -> {
-            colorMatrix.set(
-                floatArrayOf(
-                    -1f, 0f, 0f, 0f, 255f,
-                    0f, -1f, 0f, 0f, 255f,
-                    0f, 0f, -1f, 0f, 255f,
-                    0f, 0f, 0f, 1f, 0f,
-                ),
-            )
-        }
-    }
-
-    val paint =
-        Paint().apply {
-            colorFilter = ColorMatrixColorFilter(colorMatrix)
-        }
-
-    return Bitmap.createBitmap(originalBitmap.width, originalBitmap.height, Bitmap.Config.ARGB_8888).also { bitmap ->
-        val canvas = Canvas(bitmap)
-        canvas.drawBitmap(originalBitmap, 0f, 0f, paint)
-    }
-}
-
-private fun rotateImageIfRequired(
-    context: Context,
-    bitmap: Bitmap,
-    uri: Uri,
-): Bitmap {
-    val inputStream = context.contentResolver.openInputStream(uri) ?: return bitmap
-    val exif = ExifInterface(inputStream)
-    val orientation = exif.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL)
-
-    val matrix = Matrix()
-    when (orientation) {
-        ExifInterface.ORIENTATION_ROTATE_90 -> matrix.postRotate(90f)
-        ExifInterface.ORIENTATION_ROTATE_180 -> matrix.postRotate(180f)
-        ExifInterface.ORIENTATION_ROTATE_270 -> matrix.postRotate(270f)
-    }
-
-    return Bitmap.createBitmap(bitmap, 0, 0, bitmap.width, bitmap.height, matrix, true)
 }
